requirements.txt
Flask==3.1.0
Flask-WTF==1.2.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
pytz==2025.2
pillow==11.2.1
Werkzeug==3.1.3
WTForms==3.2.1
Jinja2==3.1.6
email_validator==2.2.0
SQLAlchemy==2.0.40
psycopg2==2.9.10
gunicorn==23.0.0
itsdangerous==2.2.0
MarkupSafe==3.0.2
click==8.1.8
requests==2.32.3
